# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/countries/CountryDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.countries;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class CountryDTO {
    private String name;
    private String code;
}